--- a/NetCasbin/Util/Function/IPMatchFunc.cs
+++ b/NetCasbin/Util/Function/IPMatchFunc.cs
@@ -1,17 +1,18 @@
 ﻿using System;
+using System.Net;
 using NetCasbin.Abstractions;
 
 namespace NetCasbin.Util.Function
 {
-    public class IpMatchFunc : AbstractFunction
+    public class IPMatchFunc : AbstractFunction
     {
         protected override Delegate GetFunc()
         {
-            Func<string, string, bool> call = BuiltInFunctions.IpMatch;
+            Func<string, string, bool> call = BuiltInFunctions.IPMatch;
             return call;
         }
 
-        public IpMatchFunc() : base("ipMatch")
+        public IPMatchFunc() : base("ipMatch")
         {
 
         }