close:  #85 

There are performance changes, The new version's `Mean` is nearly 6690% times shorter than the old version and the allocation is about 12228% times less.
 ```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.20201
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100-preview.8.20417.9
  [Host]     : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
  Job-ENGJLM : .NET Framework 4.8 (4.8.4210.0), X64 RyuJIT
  Job-VTJFVJ : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT

IterationCount=10  RunStrategy=Throughput  
```
Old :
|  Method |        Job |       Runtime |          key1 |           key2 |     Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |----------- |-------------- |-------------- |--------------- |---------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|
| **IPMatch** | **Job-ENGJLM** |      **.NET 4.8** | **192.168.2.123** | **192.168.2.0/24** | **25.52 μs** | **3.348 μs** | **1.992 μs** |  **1.21** |    **0.08** | **7.5989** |     **-** |     **-** |   **23.4 KB** |
| IPMatch | Job-VTJFVJ | .NET Core 3.1 | 192.168.2.123 | 192.168.2.0/24 | 21.04 μs | 1.301 μs | 0.860 μs |  1.00 |    0.00 | 6.3782 |     - |     - |  19.58 KB |
|         |            |               |               |                |          |          |          |       |         |        |       |       |           |
| **IPMatch** | **Job-ENGJLM** |      **.NET 4.8** | **192.168.2.123** |  **192.168.2.123** | **24.02 μs** | **1.850 μs** | **1.223 μs** |  **1.14** |    **0.10** | **7.4768** |     **-** |     **-** |     **23 KB** |
| IPMatch | Job-VTJFVJ | .NET Core 3.1 | 192.168.2.123 |  192.168.2.123 | 21.26 μs | 2.381 μs | 1.575 μs |  1.00 |    0.00 | 6.2866 |     - |     - |  19.29 KB |

New :
|  Method |        Job |       Runtime |          key1 |           key2 |     Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |----------- |-------------- |-------------- |--------------- |---------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|
| **IPMatch** | **Job-TYLCCQ** |      **.NET 4.8** | **192.168.2.123** | **192.168.2.0/24** | **447.8 ns** | **19.99 ns** | **11.90 ns** |  **1.41** |    **0.07** | **0.1221** |     **-** |     **-** |     **385 B** |
| IPMatch | Job-YTCUUC | .NET Core 3.1 | 192.168.2.123 | 192.168.2.0/24 | 315.5 ns | 24.02 ns | 15.89 ns |  1.00 |    0.00 | 0.0505 |     - |     - |     160 B |
|         |            |               |               |                |          |          |          |       |         |        |       |       |           |
| **IPMatch** | **Job-TYLCCQ** |      **.NET 4.8** | **192.168.2.123** |  **192.168.2.123** | **359.2 ns** | **44.25 ns** | **29.27 ns** |  **1.35** |    **0.15** | **0.0968** |     **-** |     **-** |     **305 B** |
| IPMatch | Job-YTCUUC | .NET Core 3.1 | 192.168.2.123 |  192.168.2.123 | 267.4 ns | 21.48 ns | 14.21 ns |  1.00 |    0.00 | 0.0253 |     - |     - |      80 B |