@xcaptain 2020-09-10
Very smart algorithms, looks good to me, how about also benchmark this and see which is better
https://github.com/lduchosal/ipnetwork/blob/4d931f41625d9cb56ab667f5ddc8aeda8f92006f/src/System.Net.IPNetwork/IPNetwork.cs#L1080-L1101