@Sagilio 2020-09-11
@xcaptain Here is the result:
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.20201
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100-preview.8.20417.9
  [Host]     : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
  Job-YDWRSB : .NET Framework 4.8 (4.8.4210.0), X64 RyuJIT
  Job-HEKVFF : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT

IterationCount=10  RunStrategy=Throughput  

```
|             Method |        Job |       Runtime |           ip1 |            ip2 |       Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |----------- |-------------- |-------------- |--------------- |-----------:|----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
|            **IPMatch** | **Job-YDWRSB** |      **.NET 4.8** | **192.168.2.123** | **192.168.2.0/24** |   **582.2 ns** |  **99.53 ns** |  **65.83 ns** |  **1.65** |    **0.23** | **0.1221** |     **-** |     **-** |     **385 B** |
|         IPMatchNew | Job-YDWRSB |      .NET 4.8 | 192.168.2.123 | 192.168.2.0/24 |   492.1 ns |  55.78 ns |  36.90 ns |  1.40 |    0.16 | 0.1221 |     - |     - |     385 B |
| IPMatchOfIPNetwork | Job-YDWRSB |      .NET 4.8 | 192.168.2.123 | 192.168.2.0/24 | 8,482.0 ns | 821.93 ns | 543.65 ns | 24.07 |    2.06 | 1.2512 |     - |     - |    3950 B |
|            IPMatch | Job-HEKVFF | .NET Core 3.1 | 192.168.2.123 | 192.168.2.0/24 |   354.2 ns |  46.54 ns |  30.78 ns |  1.00 |    0.00 | 0.0505 |     - |     - |     160 B |
|         IPMatchNew | Job-HEKVFF | .NET Core 3.1 | 192.168.2.123 | 192.168.2.0/24 |   362.2 ns |  39.83 ns |  26.35 ns |  1.03 |    0.13 | 0.0505 |     - |     - |     160 B |
| IPMatchOfIPNetwork | Job-HEKVFF | .NET Core 3.1 | 192.168.2.123 | 192.168.2.0/24 | 5,227.0 ns | 820.68 ns | 542.83 ns | 14.84 |    1.84 | 0.8545 |     - |     - |    2681 B |
|                    |            |               |               |                |            |           |           |       |         |        |       |       |           |
|            **IPMatch** | **Job-YDWRSB** |      **.NET 4.8** | **192.168.2.123** |  **192.168.2.123** |   **305.9 ns** |  **24.84 ns** |  **16.43 ns** |  **1.21** |    **0.10** | **0.0968** |     **-** |     **-** |     **305 B** |
|         IPMatchNew | Job-YDWRSB |      .NET 4.8 | 192.168.2.123 |  192.168.2.123 |   325.6 ns |  23.11 ns |  15.29 ns |  1.29 |    0.05 | 0.0968 |     - |     - |     305 B |
| IPMatchOfIPNetwork | Job-YDWRSB |      .NET 4.8 | 192.168.2.123 |  192.168.2.123 | 7,772.1 ns | 569.03 ns | 338.62 ns | 30.76 |    1.43 | 1.3657 |     - |     - |    4303 B |
|            IPMatch | Job-HEKVFF | .NET Core 3.1 | 192.168.2.123 |  192.168.2.123 |   252.7 ns |  13.43 ns |   8.88 ns |  1.00 |    0.00 | 0.0253 |     - |     - |      80 B |
|         IPMatchNew | Job-HEKVFF | .NET Core 3.1 | 192.168.2.123 |  192.168.2.123 |   245.7 ns |  22.94 ns |  13.65 ns |  0.97 |    0.06 | 0.0253 |     - |     - |      80 B |
| IPMatchOfIPNetwork | Job-HEKVFF | .NET Core 3.1 | 192.168.2.123 |  192.168.2.123 | 5,722.2 ns | 516.48 ns | 341.62 ns | 22.69 |    1.81 | 0.9232 |     - |     - |    2905 B |

Code:
``` csharp
public static bool IPMatchNew(string ip1, string ip2)
{
    // Same as now version

    const int ipv4Length = 4;
    const int ipv6Length = 16;
    // If IPv4 and cidrMask = 24, (ipv4Length * 8 - cidrMask) = 8,
    // -1 = 0xFFFFFFFF, (-1 << (ipv4Length * 8 - cidrMask)) = 0xFFFFFF00
    // IPAddress.NetworkToHostOrder(-1 << (ipv4Length * 8 - cidrMask)) = 0x00FFFFFF
    long cidrMaskNetWorkOrder = ip1Address.AddressFamily switch
    {
        AddressFamily.InterNetwork => BitConverter.IsLittleEndian ? ~(-1 << cidrMask) : -1 << (ipv4Length * 8 - cidrMask),
        AddressFamily.InterNetworkV6 => throw new NotImplementedException(),
        _ => throw new NotSupportedException("Unable support other address family.")
    };

    return (ip1Address.Address & cidrMaskNetWorkOrder) == (ip2Address.Address & cidrMaskNetWorkOrder);
}

public bool IPMatchOfIPNetwork(string ip1, string ip2)
{
    if (IPNetwork.TryParse(ip2, out var ipNetwork2) is false)
    {
        throw new ArgumentException($"The argument {nameof(ip1)} is not an IP address.");
    }

    if (IPAddress.TryParse(ip1, out var ipAddress1) is false)
    {
        throw new ArgumentException($"The argument {nameof(ip2)} is not an IP address.");
    }

    return ipNetwork2.Contains(ipAddress1);
}
```