@xcaptain 2020-09-11
@Sagilio Nice, your algorithm is better, go for it.